+++ winston/stats.py
@@ -6,13 +6,16 @@ import numpy
 
 from rasterio.mask import mask
 
+from shapely.geometry import mapping
+
 
 def summary(raster, geometry=None, all_touched=True, mean_only=False,
             bounds=None):
     """Return ``ST_SummaryStats`` style stats for the given raster.
 
     If ``geometry`` is provided, we mask the raster with the given geometry and
-    return the stats for the intersection.
+    return the stats for the intersection. The parameter can either be a
+    GeoJSON-like object or a Shapely geometry.
 
     If ``all_touched`` is set, we include every pixel that is touched by the
     given geometry. If set to ``False``, we only include pixels that are
@@ -45,6 +48,8 @@ def summary(raster, geometry=None, all_touched=True, mean_only=False,
 
     try:
         if geometry:
+            if not isinstance(geometry, dict):
+                geometry = mapping(geometry)
             result, _ = mask(
                 raster, [geometry], crop=True, all_touched=all_touched,
             )