+++ HISTORY.rst
@@ -2,10 +2,15 @@
 History
 =======
 
+Unreleased changes
+------------------
+
+* ``winston.stats.summary`` now accepts Shapely geometries as well as GeoJSON-like objects
+
 0.1.1 (2016-10-12)
 ------------------
 
-Minor packaging fixes.
+* Minor packaging fixes.
 
 0.1.0 (2016-10-12)
 ------------------